リファクタ：コードの名前空間とディレクトリを整理
