+++ Runtime/CityImport/AreaSelector/Display/Gizmos/AreaRectangles.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 27d0393f9f474f949e25466519a7eec9
+timeCreated: 1698733684
